feat(frontend): Proxy auth & api to backend server
Rather than referencing the API from the frontend via a URL, we assume it is on the same host as the frontend.

This is needed because we're using cookies for auth and require these to be on the same domain as  the frontend.

To make this work in development, when the frontend dev server might need to use an API on another host, we proxy the API from the frontend devserver.